better stt abstractions
