+++ main.py
@@ -26,7 +26,6 @@ if __name__ == "__main__":
     john = Chatbot(device=device, sys_prompt=call_pre_prompt)
     # john.generate_response('hello', c)
 
-
     mouth = Mouth(device=device)
     # mouth.say('Good morning! Thank you for calling Apple. My name is John, how can I assist you today?')
     mouth.say('Good morning!', ear.interrupt_listen)
