+++ stt/__init__.py
@@ -1,2 +1,3 @@
 from .stt_hf import Ear as Ear_hf
 from .stt_vosk import Ear as Ear_vosk
+
