+++ stt/stt_vosk.py
@@ -1,22 +1,19 @@
 import vosk
 import numpy as np
 import json
-from utils import record_user, record_interruption
-import re
-from vad import VoiceActivityDetection
 import torchaudio
 import torchaudio.functional as F
 import torch
+from base import BaseEar
 
 
-class Ear:
+class Ear(BaseEar):
     def __init__(self, model_path='models/vosk-model-en-us-0.22', device='cpu', silence_seconds=2):
+        super().__init__(silence_seconds)
         self.model = vosk.Model(model_path)
         self.recognizer = vosk.KaldiRecognizer(self.model, 16000)
         self.device = device
-        self.silence_seconds = silence_seconds
-        self.vad = VoiceActivityDetection()
-        self.not_interrupt_words = ['you', 'yes', 'yeah', 'hmm']
+
 
     def transcribe(self, audio):
         # if audio is a tensor convert it to numpy array
@@ -25,7 +22,6 @@ class Ear:
         audio = audio.astype(np.float64) * (1 << 15)
         audio = audio.astype(np.int16).tobytes()
         result_text = ''
-        last_partial_text = ''
         i = 0
         while True:
             data = audio[i * 12000: (i + 1) * 12000]
@@ -38,36 +34,9 @@ class Ear:
                 else:
                     result_text += json.loads(self.recognizer.Result())['text']
             else:
-                # print(self.recognizer.PartialResult())
                 pass
 
-
         return result_text
-
-    def listen(self):
-        audio = record_user(self.silence_seconds, self.vad)
-        text = self.transcribe(audio)
-        return text
-
-    def interrupt_listen(self, record_seconds=100):
-        while record_seconds > 0:
-            interruption_audio = record_interruption(self.vad, record_seconds)
-            # duration of interruption audio
-            if interruption_audio is None:
-                return False
-            else:
-                duration = len(interruption_audio) / 16_000
-                text = self.transcribe(interruption_audio)
-                # remove any punctuation using re
-                text = re.sub(r'[^\w\s]', '', text)
-                text = text.lower()
-                text = text.strip()
-                print(text)
-                if text in self.not_interrupt_words:
-                    record_seconds -= duration
-                else:
-                    return True
-
 
 if __name__ == "__main__":
     ear = Ear(model_path='../models/vosk-model-en-us-0.22')
